Moved billing scss files into views/billing/styles folder. Fixed typo in LandingPage import filepath in App.js. Fixed import filepath for App.module.scss in App.js. Fixed import filepath for styles/scss/main.scss in landing.module.scss.
