+++ src/views/billing/SubPlansGrid.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import SubPlanCard from './SubPlanCard';
-import styles from '../../scss/components/SubPlansGrid.module.scss';
+import styles from './styles/SubPlansGrid.module.scss';
 
 function SubPlansGrid(props) {
     return (
